+++ itachallenge-user/src/main/java/com/itachallenge/user/dto/SolvedDto.java
@@ -1,0 +1,15 @@
+package com.itachallenge.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class SolvedDto {
+
+    private boolean isSolved;
+
+    private int timesSolved;
+}